+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added support for API Tokens.
   See [the Authentication Guide](https://rokka.io/documentation/guides/authentication.html#using-rokka-with-a-jwt-token) for details
-- Added possibility to add comment to new membersips
+- Added possibility to add a comment to new memberships
 
 # 3.6.0 [01-01-2022]